offloaded production traefik to external instance
This is to allow a shared traefik instance to proxy multiple services.
During development the authentication will be overwritten in the development file
and it will only be extended with an extra basicauth in production.
This is done to allow pushes from pipelines as alternative to non-interactive oauth
while skipping the requirement for oauth on localhost (fallback to basicauth here)

Signed-off-by: Manuel Zedel <manuel.zedel@northern.tech>